major bug fixes

+++ backend/app/models/food.py
@@ -2,7 +2,7 @@
 Food models - BCNF normalized
 Separated into Food (catalog) and FoodItem (user's food entries)
 """
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, Text
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Text, Numeric
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.core.database import Base
@@ -15,13 +15,13 @@ class Food(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(255), nullable=False, index=True)
     description = Column(Text)
-    calories_per_100g = Column(Float, nullable=False)
-    protein_per_100g = Column(Float, nullable=False)
-    carbs_per_100g = Column(Float, nullable=False)
-    fats_per_100g = Column(Float, nullable=False)
-    fiber_per_100g = Column(Float, default=0.0)
-    sugar_per_100g = Column(Float, default=0.0)
-    sodium_per_100g = Column(Float, default=0.0)
+    calories_per_100g = Column(Numeric(10, 2), nullable=False)
+    protein_per_100g = Column(Numeric(10, 2), nullable=False)
+    carbs_per_100g = Column(Numeric(10, 2), nullable=False)
+    fats_per_100g = Column(Numeric(10, 2), nullable=False)
+    fiber_per_100g = Column(Numeric(10, 2), default=0.0)
+    sugar_per_100g = Column(Numeric(10, 2), default=0.0)
+    sodium_per_100g = Column(Numeric(10, 2), default=0.0)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     
     # Relationships
@@ -35,7 +35,7 @@ class FoodItem(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     food_id = Column(Integer, ForeignKey("foods.id"), nullable=False)
-    quantity_g = Column(Float, nullable=False)
+    quantity_g = Column(Numeric(10, 2), nullable=False)
     custom_name = Column(String(255))  # User can rename
     notes = Column(Text)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
