+++ backend/app/models/preference.py
@@ -1,7 +1,7 @@
 """
 User preference models - BCNF normalized
 """
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, Float
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.core.database import Base
@@ -13,10 +13,10 @@ class UserPreference(Base):
     
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False, unique=True)
-    target_calories = Column(Float)
-    target_protein = Column(Float)
-    target_carbs = Column(Float)
-    target_fats = Column(Float)
+    target_calories = Column(Numeric(10, 2))
+    target_protein = Column(Numeric(10, 2))
+    target_carbs = Column(Numeric(10, 2))
+    target_fats = Column(Numeric(10, 2))
     preferred_meal_times = Column(String(500))  # JSON string
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
